// src/types/cssClassEnums.ts
export enum TOOLTIP_VARIANT {
  TOOLTIP_SMALL = 'tooltip-small',
  TOOLTIP_MEDIUM = 'tooltip-medium',
}

export enum TOOLTIP_TYPE {
  TOOLTIP_SINGLE_LINE = 'tooltip-single-line',
  TOOLTIP_MULTI_LINE = 'tooltip-multi-line',
}

export enum TOOLTIP_PLACEMENT {
  TOOLTIP_BOTTOM = 'tooltip-bottom',
  TOOLTIP_BOTTOM_START = 'tooltip-bottom-start',
  TOOLTIP_BOTTOM_END = 'tooltip-bottom-end',
  TOOLTIP_TOP = 'tooltip-top',
  TOOLTIP_TOP_START = 'tooltip-top-start',
  TOOLTIP_TOP_END = 'tooltip-top-end',
  TOOLTIP_RIGHT = 'tooltip-right',
  TOOLTIP_RIGHT_START = 'tooltip-right-start',
  TOOLTIP_RIGHT_END = 'tooltip-right-end',
  TOOLTIP_LEFT = 'tooltip-left',
  TOOLTIP_LEFT_START = 'tooltip-left-start',
  TOOLTIP_LEFT_END = 'tooltip-left-end',
}